More minor typecasts
git-svn-id: https://spring.clan-sy.com/svn/spring/trunk@271 37977431-3df6-0310-b722-df95706aa16b

+++ rts/BubbleProjectile.cpp
@@ -14,7 +14,7 @@ CBubbleProjectile::CBubbleProjectile(float3 pos,float3 speed,float ttl,float sta
 	startSize(startSize),
 	size(startSize*0.4),
 	sizeExpansion(sizeExpansion),
-	ttl(ttl)
+	ttl((int)ttl)
 {
 	checkCol=false;
 }
